requirements.txt
Flask==3.0.3
Pillow==10.4.0
pytesseract==0.3.13
opencv-python==4.10.0.84
pyzbar==0.1.9
numpy==2.0.1
gunicorn==21.2.0